requirements.txt
# Core Framework
fastapi[standard]
pydantic-settings
httpx

# Database Layer
mariadb
sqlalchemy
alembic
python-dotenv

# AI Model
torch
transformers
sentencepiece

# Testing
pytest
pytest-asyncio

# Security
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart